requirements.txt
annotated-types==0.7.0
anyio==4.7.0
apturl==0.5.2
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
bcrypt==3.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
Brlapi==0.8.3
certifi==2020.6.20
cffi==1.17.1
channels==4.0.0
chardet==4.0.0
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.4
command-not-found==0.3
constantly==23.10.4
cryptography==44.0.1
cupshelpers==1.0
daphne==4.0.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.9
distro==1.7.0
distro-info==1.1+ubuntu0.2
dj-rest-auth==4.0.0
Django==5.0.2
django-allauth==0.52.0
django-cors-headers==4.3.1
django-debug-toolbar==5.0.1
django-extensions==3.2.3
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
duplicity==0.8.21
exceptiongroup==1.2.2
Faker==37.0.2
fasteners==0.14.1
filelock==3.16.1
Flask==3.1.0
future==0.18.2
gpg==1.16.0
gunicorn==21.2.0
gyp==0.1
h11==0.14.0
hiredis==3.1.0
httpcore==1.0.7
httplib2==0.20.2
httpx==0.28.1
huepy==1.2.1
hyperlink==21.0.0
idna==3.3
imageio==2.36.0
imageio-ffmpeg==0.5.1
importlib-metadata==4.6.4
incremental==24.7.2
iniconfig==2.0.0
instabot==0.117.0
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.5
jiter==0.8.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==3.0.2
mock==5.1.0
monotonic==1.6
more-itertools==8.10.0
moviepy==1.0.3
mysql-connector-python==9.1.0
mysqlclient==2.1.1
natsort==8.0.2
netifaces==0.11.0
numpy==2.1.3
oauthlib==3.2.0
olefile==0.46
openai==1.57.0
packaging==24.1
paramiko==2.9.3
pexpect==4.8.0
pillow==10.2.0
platformdirs==4.3.6
pluggy==1.5.0
proglog==0.1.10
protobuf==3.12.4
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
pydantic==2.10.3
pydantic_core==2.27.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==4.10.1
PyMySQL==1.1.1
PyNaCl==1.5.0
pyOpenSSL==25.0.0
pyparsing==2.4.7
pyRFC3339==1.1
PySocks==1.7.1
pytest==8.3.3
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
qrcode==8.0
redis==5.2.1
reportlab==3.6.8
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
responses==0.25.3
schedule==1.2.2
SecretStorage==3.3.1
service-identity==24.2.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.11.0
sqlparse==0.5.3
systemd-python==234
tomli==2.0.2
tqdm==4.67.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
uv==0.5.14
virtualenv==20.28.0
wadllib==1.3.6
Werkzeug==3.1.3
wikipedia==1.4.0
wxPython==4.0.7
xdg==5
xkit==0.0.0
zipp==1.0.0
zope.interface==7.2
